requirements.txt
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
dominate==2.4.0
Flask==1.1.1
Flask-Assets==0.12
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.8
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
oauthlib==3.1.0
pusher==2.1.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyScss==1.3.5
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.16.0
